+++ src/common/io/io.h
@@ -120,4 +120,14 @@ static inline bool wrapClose(FFNativeFD* pfd)
 }
 #define FF_AUTO_CLOSE_FD __attribute__((__cleanup__(wrapClose)))
 
+static inline bool wrapFclose(FILE** pfile)
+{
+    assert(pfile);
+    if (!*pfile)
+        return false;
+    fclose(*pfile);
+    return true;
+}
+#define FF_AUTO_CLOSE_FILE __attribute__((__cleanup__(wrapFclose)))
+
 #endif // FF_INCLUDED_common_io_io